% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/n1simulator.R
\name{n1_simulate_functions}
\alias{n1_simulate_functions}
\title{Simulate N-of-1 study and return results as continuous functions of time.}
\usage{
n1_simulate_functions(t_change_vec, B_func, E_vec, T_mat, tau0_vec, tau1_vec,
  tc_outcome, sigma_Z, dt, X_initial_vec = NA, Z_initial = NA)
}
\arguments{
\item{t_change_vec}{Sorted vector of simulation \emph{change} times.}
}
\description{
Simulate N-of-1 study and return results as continuous functions of time.
}
